+++ acme/resources/DEPR.py
@@ -1,7 +1,7 @@
 #
 #	DEPR.py
 #
-#	(c) 2022 by Andreas Kraft
+#	(c) 2023 by Andreas Kraft
 #	License: BSD 3-Clause License. See the LICENSE file for further details.
 #
 #	ResourceType: Dependency